requirements.txt
flask==3.1.1
pymongo==4.13.2
python-dotenv==1.0.0
requests==2.32.4
Werkzeug==3.1.3
bson==0.5.10
deep-translator==1.11.4
langdetect==1.0.9
beautifulsoup4==4.13.4
dnspython==2.7.0  # Required for MongoDB connection
python-dotenv==1.0.0
certifi==2025.7.9  # Required for HTTPS requests
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
